// frontend/src/types/Metrics.ts
/** @format */

export type Metrics = {
  CVSSV4_version?: string;
  CVSSV4_vector_string?: string;
  CVSSV4_base_score?: number;
  CVSSV4_base_severity?: string;
  CVSSV3_1_version?: string;
  CVSSV3_1_vector_string?: string;
  CVSSV3_1_base_score?: number;
  CVSSV3_1_base_severity?: string;
  CVSSV3_version?: string;
  CVSSV3_vector_string?: string;
  CVSSV3_base_score?: number;
  CVSSV3_base_severity?: string;
  CVSSV2_version?: string;
  CVSSV2_vector_string?: string;
  CVSSV2_base_score?: number;
  CVSSV2_base_severity?: string;
};